tests assistant added
